Refactor: move activity-related logic in transport to abstract TransportActivityManager. DefaultTransportService now extends TransportActivityManager.
